Search for department and entities fixes

+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement.Contracts/Documents/DocumentPdfDetails.cs
@@ -4,7 +4,7 @@ namespace DigitNow.Domain.DocumentManagement.Contracts.Documents
 {
     public class DocumentPdfDetails
     {
-        public int RegistrationNumber { get; set; }
+        public long RegistrationNumber { get; set; }
         public DateTime RegistrationDate { get; set; }
         public string IssuerName { get; set; }
         public string DocumentType { get; set; }
